Update src/backbone-validation.js
When using a model/view binder in a form, the validator is called after
a form field is edited because the binder sets the new value to the model.
This could be annoying if you just want to fill out the form out and 
then only validate on save. There could be an option to only validate
on save, or there could be an option to ignore change sources. The
changeSource variable is given to us from the setOptions argument in
the validate function. We could use this to determine whether or not
to validate.