Added ability to preserve intermediate levels during flatten operation.
Currently this is set up as an extra function argument, but unfortunately all callers are using it. I set it up this way in the hopes of being able to use it only sometimes. As people play around with it, I hope to determine whether or not all flattening needs to have intermediate levels preserved or not; if so, I will remove the parameter, and if not, I will modify the calls that don't need it.