Fixing and reorganizing docco to actually work.
Now that a CSS file seems to be required, I have decided to copy the CSS file and use it as input. Rather than having both docco.jst and docco.css at the top level, I decided to move them to a new subfolder "docco" and have the gruntfile pull them in from there.