Fixes #282: Ensure options are passed into validator calling context to allow for overriding of labelFormatter.
Unfortunately this required changing how we handle calls from one validator to another (default) validator: _.bind now gets in the way because the context object was created without an options object. Calling other validators with  will allow the options to continue along, but in the meantime there is a new hack to avoid "this.formatLabel is not a function" errors.

We need to tweak how all of this is handled. Either we need to start treating Validation mixins as proper instances, or we need to overhaul the calling-other-validators functionality to make it more sensible. As it is, calling other validators still works, but #282 is not fixed in that case without .